+++ sql/3.5/orcadbup_teikiprt_3.5.0-1.sql
@@ -17,7 +17,7 @@ drop table tbl_teikiprt ;
 
 
 CREATE TABLE "tbl_teikiprt" (
-	"hospnum" integer NOT NULL,
+	"hospnum" numeric(2,0) NOT NULL,
 	"btunum" character(2) NOT NULL,
 	"denpnum" numeric(7,0) NOT NULL,
 	"ptid" numeric(10,0),
